+++ src/ngx_http_haskell_load.c
@@ -174,6 +174,16 @@ ngx_http_haskell_load(ngx_cycle_t *cycle)
         goto dlclose_and_exit;
     }
 
+    mcf->set_upstream_main_conf_ptr = (void (*)(HsPtr)) dlsym(mcf->dl_handle,
+                                            "ngxExportSetUpstreamMainConfPtr");
+    dl_error = dlerror();
+    if (dl_error != NULL) {
+        ngx_log_error(NGX_LOG_EMERG, cycle->log, 0,
+                      "failed to load function "
+                      "\"ngxExportSetUpstreamMainConfPtr\": %s", dl_error);
+        goto dlclose_and_exit;
+    }
+
     install_signal_handler = (void (*)(void)) dlsym(mcf->dl_handle,
                                             "ngxExportInstallSignalHandler");
     dl_error = dlerror();